refactor(decisions): move business logic from routes to decision layer (#284)
Extract 7 helper functions from routes/decisions.ts into
decision/route-helpers.ts. Routes layer now only handles HTTP
concerns and delegates business logic to the decision layer.
